WIP: Making timeseries key more robust and less visible
The idea of timeseries keys is really only relevant to the ClickHouse
database, allowing us to correlate multiple records after normalizing
the incoming data. Most oximeter consumers don't care and shouldn't know
about this. This moves the keys out of the Target/Metric traits, and
into a concept only the DB understands.

This also makes the keys more robust. Previously, these were strings
built by concatenating the field values with `":"` as a separator. This
worked, but it obviously means we can't have fields with the `:`
character in them. ClickHouse supports robust array types, so this
commit uses an array of strings for the keys, rather than a single
concatenated string.

The timeseries key is no longer returned in the
`oximeter_db::model::Timeseries` type, since it's not relevant and
duplicative with the field values. This also makes a lot of types inside
the models module private that were unnecessarily public.